add sweep

+++ assignment1-basics/cs336_basics/experiment/hyperparam_sweep.py
@@ -6,7 +6,7 @@ from cs336_basics import my_train
 def main():
     parser = argparse.ArgumentParser()
     parser.add_argument("command", choices=["lr", "batch"])
-    parser.add_argument("values", nargs="+", type=int)
+    parser.add_argument("values", nargs="+", type=float)
     args = parser.parse_args()
 
     config = my_train.get_config()
@@ -22,7 +22,7 @@ def main():
     elif args.command == "batch":
         for val in args.values:
             config.train.wandb.name = f"sweep_{args.command}_{val}"
-            config.train.batch_size = val
+            config.train.batch_size = int(val)
             print(f"Running with batch_size={val}")
             my_train.main(config)
